+++ sdk/communication/AzureCommunicationCommon/CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Features Added
 - Optimization added: When the proactive refreshing is enabled and the token refresher fails to provide a token that's not about to expire soon, the subsequent refresh attempts will be scheduled for when the token reaches half of its remaining lifetime until a token with long enough validity (>10 minutes) is obtained.
 - Added `cancel()` to `CommunicationTokenCredential` that cancels any internal auto-refresh operation. An instance of `CommunicationTokenCredential` cannot be reused once it has been canceled, otherwise the error will be returned.
+- Added support for a new communication identifier `MicrosoftBotIdentifier`.
 
 ## 1.1.1 (2022-11-17)
 ### Bugs fixed